chore: rename to ScriptableGraph & reorganise
