+++ Runtime/Scripting/Graph/CareerGraph.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ccad7390709409a41aac59bd38f4e1a1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
