--- Runtime/Scripting/Graph/ContractGraph.cs
@@ -1,10 +1,0 @@
-using UnityEngine;
-
-namespace ContractGraph
-{
-    [CreateAssetMenu(fileName = "NewContractGraph", menuName = "Contracts/ContractGraph")]
-    public class ContractGraph : ContractBaseGraph
-    {
-
-    }
-}
