+++ Runtime/Scripting/Graph/ContractGraph/ContractGraph.cs
@@ -1,0 +1,10 @@
+using UnityEngine;
+
+namespace Contracts.Scripting.Graph
+{
+    [CreateAssetMenu(fileName = "NewContractGraph", menuName = "Contracts/ContractGraph")]
+    public class ContractGraph : ScriptableGraph
+    {
+
+    }
+}
