+++ Runtime/Scripting/Graph/ContractGraph/Nodes/ConditionNode.cs
@@ -1,0 +1,21 @@
+using UnityEditor.Experimental.GraphView;
+using UnityEngine;
+
+namespace Contracts.Scripting.Graph
+{
+    [NodeMenu("Condition")]
+    public class ConditionNode : ScriptableGraphNode
+    {
+        [NodeOutput("Satisfied", Port.Capacity.Single)]
+        public ConditionNode Satisfied;
+
+        [NodeOutput("Dissatisfied", Port.Capacity.Single)]
+        public ConditionNode Dissatisfied;
+
+        public ConditionNode() : base("Condition", new Color(0.3f, 0.3f, 0.6f))
+        {
+            Satisfied = this;
+            Dissatisfied = this;
+        }
+    }
+}
