--- Runtime/Scripting/Graph/ContractGraphNode.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: bdcd99df3f692d54fbeaa6d16eebdfc7
