--- Runtime/Scripting/Graph/ContractGraphView.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 63f596870aad0df4699aa5ac3758abdc
