--- Runtime/Scripting/Graph/EdgeConnectorListener.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: c9cd65a4b7da4014ca1bd21521bd4e39
