--- Runtime/Scripting/Graph/EmptyGraph.cs
@@ -1,19 +1,0 @@
-using UnityEngine;
-
-namespace Contracts
-{
-    public class ContractGraph : MonoBehaviour
-    {
-        // Start is called once before the first execution of Update after the MonoBehaviour is created
-        void Start()
-        {
-        
-        }
-
-        // Update is called once per frame
-        void Update()
-        {
-        
-        }
-    }
-}
