--- Runtime/Scripting/Graph/EmptyGraph.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: d331d40fd765fee4991a38623ab55567
