+++ Runtime/Scripting/Graph/ScriptableGraphNode.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: b42e54f55692dff40912b405355bd806
