--- Runtime/Scripting/Graph/USS/ContractGraphView.uss
@@ -1,10 +1,0 @@
-ContractGraphView
-{
-    flex-grow: 1;
-}
-#Grid
-{
-    --spacing: 32;
-    --line-color: rgb(50,50,50);
-    --thick-line-color: rgb(50,50,50);
-}
