+++ Runtime/Scripting/Graph/USS/ScriptableGraphView.uss
@@ -1,0 +1,10 @@
+ContractGraphView
+{
+    flex-grow: 1;
+}
+#Grid
+{
+    --spacing: 32;
+    --line-color: rgb(50,50,50);
+    --thick-line-color: rgb(50,50,50);
+}
